Core tweaks
- use the new `Call.append` in the listener's channel create handler
- move `get_event_time` to the `utils` mod
- always return the app client id from `Client.load_app` even if the app
  is already registered
